refactor: move `ConnectionTrigger`
